Santosh

+++ Santosh/Employee.java
@@ -3,5 +3,5 @@ package myemp;
 public abstract class Employee{
     public int emp_id;
     public String full_name;
-    public int salary;
+    public double salary;
 }
